breaks out the inputs into separate stories
