+++ react/src/base/inputs/Checkbox.stories.js
@@ -1,0 +1,51 @@
+import React from 'react';
+import SprkSelectionInput from './SprkSelectionInput/SprkSelectionInput';
+
+export default {
+  title: 'Components/Input/Checkbox',
+  decorators: [
+    story => <div className="sprk-o-Box">{story()}</div>
+  ],
+  component: SprkSelectionInput,
+  parameters: {
+    jest: [
+      'SprkErrorContainer',
+      'SprkInputIconCheck',
+    ],
+    info: `
+##### For design and usage information check out the [documentation.](https://spark-docs.netlify.com/using-spark/components/input)
+    `,
+  },
+};
+
+export const checkbox = () => (
+  <SprkSelectionInput
+    groupLabel="Checkbox Input"
+    choices={[
+      {
+        label: 'Checkbox Item 1',
+        name: 'check[]',
+        value: 'item-1',
+      },
+      {
+        label: 'Checkbox Item 2',
+        name: 'check[]',
+        value: 'item-2',
+      },
+      {
+        label: 'Checkbox Item 3',
+        name: 'check[]',
+        value: 'item-2',
+      },
+    ]}
+    variant="checkbox"
+  />
+);
+
+checkbox.story = {
+  parameters: {
+    jest: [
+      'SprkSelectionInput',
+    ]
+  },
+};
