+++ react/src/base/inputs/Percentage.stories.js
@@ -1,0 +1,38 @@
+import React from 'react';
+import SprkTextInput from './SprkTextInput/SprkTextInput';
+
+export default {
+  title: 'Components/Input/Percentage',
+  decorators: [
+    story => <div className="sprk-o-Box">{story()}</div>
+  ],
+  component: SprkTextInput,
+  parameters: {
+    jest: [
+      'SprkErrorContainer',
+      'SprkInputIconCheck',
+    ],
+    info: `
+##### For design and usage information check out the [documentation.](https://spark-docs.netlify.com/using-spark/components/input)
+    `,
+  },
+};
+
+export const percentageInput = () => (
+  <SprkTextInput
+    label="Percentage"
+    iconRight
+    leadingIcon="percent"
+    narrowWidth
+    name="percentage"
+    type="tel"
+  />
+);
+
+percentageInput.story = {
+  parameters: {
+    jest: [
+      'SprkTextInput',
+    ]
+  },
+};
