+++ react/src/base/inputs/Phone.stories.js
@@ -1,0 +1,36 @@
+import React from 'react';
+import SprkTextInput from './SprkTextInput/SprkTextInput';
+
+export default {
+  title: 'Components/Input/Phone',
+  decorators: [
+    story => <div className="sprk-o-Box">{story()}</div>
+  ],
+  component: SprkTextInput,
+  parameters: {
+    jest: [
+      'SprkErrorContainer',
+      'SprkInputIconCheck',
+    ],
+    info: `
+##### For design and usage information check out the [documentation.](https://spark-docs.netlify.com/using-spark/components/input)
+    `,
+  },
+};
+
+export const phoneInput = () => (
+  <SprkTextInput
+    label="Phone Number"
+    name="phone"
+    placeholder="(000) 000-0000"
+    errorMessage="Incorrect phone number."
+  />
+);
+
+phoneInput.story = {
+  parameters: {
+    jest: [
+      'SprkTextInput',
+    ]
+  },
+};
