+++ react/src/base/inputs/SSN.stories.js
@@ -1,0 +1,36 @@
+import React from 'react';
+import SprkRevealInput from './SprkRevealInput/SprkRevealInput';
+
+export default {
+  title: 'Components/Input/SSN',
+  decorators: [
+    story => <div className="sprk-o-Box">{story()}</div>
+  ],
+  component: SprkRevealInput,
+  parameters: {
+    jest: [
+      'SprkErrorContainer',
+      'SprkInputIconCheck',
+    ],
+    info: `
+##### For design and usage information check out the [documentation.](https://spark-docs.netlify.com/using-spark/components/input)
+    `,
+  },
+};
+
+export const ssnInput = () => (
+  <SprkRevealInput
+    label="Social Security Number"
+    toggleLabel="Show SSN"
+    name="ssn"
+  />
+);
+
+ssnInput.story = {
+  name: 'SSN Input',
+  parameters: {
+    jest: [
+      'SprkRevealInput',
+    ]
+  },
+};
