+++ react/src/base/inputs/Search.stories.js
@@ -1,0 +1,48 @@
+import React from 'react';
+import SprkTextInput from './SprkTextInput/SprkTextInput';
+
+export default {
+  title: 'Components/Input/Search',
+  decorators: [
+    story => <div className="sprk-o-Box">{story()}</div>
+  ],
+  component: SprkTextInput,
+  parameters: {
+    jest: [
+      'SprkErrorContainer',
+      'SprkInputIconCheck',
+    ],
+    info: `
+##### For design and usage information check out the [documentation.](https://spark-docs.netlify.com/using-spark/components/input)
+    `,
+  },
+};
+
+export const searchInput = () => (
+  <SprkTextInput label="Search" name="Search" placeholder="Search" />
+);
+
+searchInput.story = {
+  parameters: {
+    jest: [
+      'SprkTextInput',
+    ]
+  }
+};
+
+export const inlineSearchInput = () => (
+  <SprkTextInput
+    leadingIcon="search"
+    hiddenLabel
+    name="InlineSearch"
+    placeholder="Search"
+  />
+);
+
+inlineSearchInput.story = {
+  parameters: {
+    jest: [
+      'SprkTextInput',
+    ]
+  },
+};
