+++ react/src/base/inputs/Select.stories.js
@@ -1,0 +1,67 @@
+import React from 'react';
+import SprkSelectionInput from './SprkSelectionInput/SprkSelectionInput';
+import SprkTextInput from './SprkTextInput/SprkTextInput';
+
+export default {
+  title: 'Components/Input/Select',
+  decorators: [
+    story => <div className="sprk-o-Box">{story()}</div>
+  ],
+  component: SprkSelectionInput,
+  parameters: {
+    jest: [
+      'SprkErrorContainer',
+      'SprkInputIconCheck',
+    ],
+    info: `
+##### For design and usage information check out the [documentation.](https://spark-docs.netlify.com/using-spark/components/input)
+    `,
+  },
+};
+
+export const selectBox = () => (
+  <SprkSelectionInput
+    choices={[
+      {
+        label: 'Option 1',
+        value: 'option-1',
+      },
+      {
+        label: 'Option 2',
+        value: 'option-2',
+      },
+      {
+        label: 'Option 3',
+        value: 'option-3',
+      },
+      {
+        label: 'Grouped Options',
+        options: [
+          {
+            label: 'Grouped Option 1',
+            value: 'grouped-option-1',
+          },
+          {
+            label: 'Grouped Option 2',
+            value: 'grouped-option-2',
+          },
+          {
+            label: 'Grouped Option 3',
+            value: 'grouped-option-3',
+          },
+        ],
+      },
+    ]}
+    name="name"
+    variant="select"
+    label="Select Box Label"
+  />
+);
+
+selectBox.story = {
+  parameters: {
+    jest: [
+      'SprkSelectionInput',
+    ]
+  },
+};
