+++ react/src/base/inputs/Text.stories.js
@@ -1,0 +1,37 @@
+import React from 'react';
+import SprkTextInput from './SprkTextInput/SprkTextInput';
+
+export default {
+  title: 'Components/Input/Text',
+  decorators: [
+    story => <div className="sprk-o-Box">{story()}</div>
+  ],
+  component: SprkTextInput,
+  parameters: {
+    jest: [
+      'SprkErrorContainer',
+      'SprkInputIconCheck',
+    ],
+    info: `
+##### For design and usage information check out the [documentation.](https://spark-docs.netlify.com/using-spark/components/input)
+    `,
+  },
+};
+
+export const textInput = () => (
+  <SprkTextInput
+    label="Text Input Label"
+    name="text-input-label"
+    valid={true}
+    disabled={false}
+    errorMessage="There is an error on this field."
+  />
+);
+
+textInput.story = {
+  parameters: {
+    jest: [
+      'SprkTextInput',
+    ]
+  },
+};
