+++ react/src/base/inputs/Textarea.stories.js
@@ -1,0 +1,31 @@
+import React from 'react';
+import SprkTextInput from './SprkTextInput/SprkTextInput';
+
+export default {
+  title: 'Components/Input/Textarea',
+  decorators: [
+    story => <div className="sprk-o-Box">{story()}</div>
+  ],
+  component: SprkTextInput,
+  parameters: {
+    jest: [
+      'SprkErrorContainer',
+      'SprkInputIconCheck',
+    ],
+    info: `
+##### For design and usage information check out the [documentation.](https://spark-docs.netlify.com/using-spark/components/input)
+    `,
+  },
+};
+
+export const textarea = () => (
+  <SprkTextInput label="Description" name="description" type="textarea" />
+);
+
+textarea.story = {
+  parameters: {
+    jest: [
+      'SprkTextInput',
+    ]
+  },
+};
